Tried many different methods here.
In the end a depth first search and a-star can be used. Also I left some optimisation methods unused.
Very important here is to limit the movement options as much as possible. i.e. only consider moving to a room if its empty and or only right units are in there. If moving a unit to a room only consider the lowest spot. Remove all inter corridor movements. Then for the heuristics man-hatten distance alone does not work, you also need to consider the traffic and estimate required cleanup energy. Finally make sure the estimates are not a lot larger that the real costs (otherwise its very fast but also doesn't consider better options).